setup *.gradle files
