+++ app/build.gradle
@@ -5,18 +5,15 @@ plugins {
 
 android {
     namespace 'com.konovalov.draggable.coin.view'
-    compileSdk 33
+    compileSdk compile_sdk_version
 
     defaultConfig {
         applicationId "com.konovalov.draggable.coin.view"
-        minSdk 29
-        targetSdk 33
+        minSdk min_sdk_version
+        targetSdk target_sdk_version
+        consumerProguardFiles 'consumer-rules.pro'
         versionCode 1
         versionName "1.0"
-
-        vectorDrawables {
-            useSupportLibrary true
-        }
     }
 
     buildTypes {
@@ -32,21 +29,13 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
-    buildFeatures {
-        viewBinding true
-        compose true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion "1.4.3"
-    }
-    packagingOptions {
-        resources {
-            excludes += '/META-INF/{AL2.0,LGPL2.1}'
-        }
-    }
 }
 
 dependencies {
+    implementation (
+            project(':features:compose'),
+            project(':features:xml'),
+    )
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.0'
@@ -55,10 +44,4 @@ dependencies {
     implementation 'androidx.navigation:navigation-fragment-ktx:2.4.1'
     implementation 'androidx.navigation:navigation-ui-ktx:2.4.1'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
-    implementation 'androidx.activity:activity-compose:1.3.1'
-    implementation "androidx.compose.ui:ui:$compose_ui_version"
-    implementation "androidx.compose.ui:ui-tooling-preview:$compose_ui_version"
-    implementation 'androidx.compose.material:material:1.2.0'
-    debugImplementation "androidx.compose.ui:ui-tooling:$compose_ui_version"
-    debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_ui_version"
 }
