+++ app/src/main/AndroidManifest.xml
@@ -12,11 +12,6 @@
         android:theme="@style/Theme.DraggableScaredView"
         tools:targetApi="31">
         <activity
-            android:name=".ComposeActivity"
-            android:exported="false"
-            android:label="@string/title_activity_compose"
-            android:theme="@style/Theme.DraggableScaredView" />
-        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:theme="@style/Theme.DraggableScaredView">
