+++ build.gradle
@@ -1,5 +1,8 @@
 buildscript {
     ext {
+        min_sdk_version = 26
+        target_sdk_version = 33
+        compile_sdk_version = 33
         compose_ui_version = '1.2.0'
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
