+++ features/compose/src/main/res/drawable/coin.xml
