+++ features/compose/src/main/res/layout/draggable_item.xml
