+++ features/compose/src/main/res/layout/fragment_first.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.konovalov.draggable.coin.view.CoinView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.konovalov.draggable.coin.view.xml.CoinView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_rv"
