+++ features/xml/build.gradle
@@ -1,11 +1,19 @@
 apply from: "$rootDir/core/common-build.gradle"
 
-dependencies {
+android {
+    namespace 'com.konovalov.draggable.coin.view.xml'
 
+    buildFeatures {
+        viewBinding true
+    }
+}
+
+dependencies {
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.8.0'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.4.1'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.4.1'
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
 }
